+++ src/test/java/BuyTshirtProcessTest.java
@@ -1,3 +1,4 @@
+import com.saucelabs.saucerest.SauceREST;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import org.junit.Rule;
 import org.junit.rules.TestRule;
@@ -19,6 +20,8 @@ import java.net.MalformedURLException;
  */
 public class BuyTshirtProcessTest {
   private SaucedomoLogInPage logIn;
+  private SauceREST sauceClient;
+  private String sessionId;
 
 
 
@@ -28,6 +31,8 @@ public class BuyTshirtProcessTest {
   @BeforeMethod
   public void setup() throws MalformedURLException {
     logIn = new SaucedomoLogInPage();
+    sauceClient = WebDriverContainer.getSauceClient();
+    sessionId = WebDriverContainer.getSessionId();
   }
 
   @Test(description = "This test is in charge of navigating through the webpage and"
@@ -48,5 +53,22 @@ public class BuyTshirtProcessTest {
   public void closeDriver() throws MalformedURLException {
     WebDriverContainer.getInstance().quit();
   }
+
+  @Rule
+  public TestRule watcher = new TestWatcher() {
+    @Override
+    protected void succeeded(Description description) {
+      if (sessionId!=null && sauceClient!=null) {
+        sauceClient.jobPassed(sessionId);
+      }
+    }
+
+    @Override
+    protected void failed(Throwable e, Description description) {
+      if (sessionId!=null && sauceClient!=null) {
+        sauceClient.jobFailed(sessionId);
+      }
+    }
+  };
 
 }
